Changed Github pipeline to fetch the latest version tag from the box2d repo and build that instead of main. This means we don't end up with partial new API support, which can be complicated by ongoing changes to the API.
Made sure everything is compatible with Box2D 3.0.0, and ready to upgrade to 3.1.0 when the time is right.

Removed all Get* and Set* where they can be replaced with properties

Added a unit test which fails if the built Box2D version is not 3.0.0 - this will ensure we never deploy a Box2D 3.0.0 API with a Box2D 3.1.0 binary.

Added MacOS (non-critical) to the publish pipeline